requirements.txt
mediapipe
opencv-python
ffmpeg-python